requirements.txt
selenium==4.10.0
Pillow==9.5.0
requests==2.31.0
tqdm==4.65.0